requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pymongo
pydantic
python-jose 